Support for different dalle models
